requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
html5lib==1.0.1
Pillow==7.1.2
psycopg2-binary==2.8.5
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.42
six==1.14.0
sqlparse==0.3.1
webencodings==0.5.1
xhtml2pdf==0.2.4
